more robust way of extracting dlls / bwta data from classpath resources
this is important, as if you want to run the TestBot in the bwmirror project
directly from an IDE, the dlls and bwta data files are not being extracted
from a JAR file and the extraction would fail. this method is a fair bit more
robust and will automatically handle those differences.

in addition, using System.load instead of System.loadLibrary to load
the bwapi_bridge native library. this allows us to stop doing that hacky
reflection thing to reset the java.library.path property and just pass in
an absolute path to bwapi_bridge.dll